Added Timer demo
